+++ addons/hr_holidays/__init__.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-import hr_department
-import hr_holidays
+import models
 import report
 import wizard
