+++ addons/hr_holidays/__openerp__.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-
 {
     'name': 'Leave Management',
     'version': '1.5',
@@ -17,7 +16,7 @@ This application controls the leave schedule of your company. It allows employee
 
 You can configure several kinds of leaves (sickness, paid days, ...) and allocate leaves to an employee or department quickly using leave allocation. An employee can also make a request for more days off by making a new Leave allocation. It will increase the total of available days for that leave type (if the request is accepted).
 
-You can keep track of leaves in different ways by following reports: 
+You can keep track of leaves in different ways by following reports:
 
 * Leaves Summary
 * Leaves by Department
@@ -28,25 +27,32 @@ A synchronization with an internal agenda (Meetings of the CRM module) is also p
     'depends': ['hr', 'calendar', 'resource', 'product'],
     'data': [
         'data/report_paperformat.xml',
+        'data/hr_holidays_data.xml',
+        'data/hr_holidays_workflow.xml',
+
         'security/ir.model.access.csv',
-        'security/ir_rule.xml',
-        'hr_holidays_workflow.xml',
-        'hr_holidays_view.xml',
-        'hr_holidays_data.xml',
-        'hr_holidays_report.xml',
-        'views/report_hr_holidays_summary.xml',
-        'report/hr_holidays_report_view.xml',
-        'report/available_holidays_view.xml',
-        'wizard/hr_holidays_summary_department_view.xml',
-        'wizard/hr_holidays_summary_employees_view.xml',
-        'hr_dashboard.xml',
-        ],
-    'demo': ['hr_holidays_demo.xml',],
+        'security/hr_holidays_security.xml',
+
+        'views/hr_views.xml',
+        'views/hr_holidays_views.xml',
+
+        'wizard/hr_holidays_summary_department_views.xml',
+        'wizard/hr_holidays_summary_employees_views.xml',
+
+        'report/hr_holidays_templates.xml',
+        'report/hr_holidays_reports.xml',
+        'report/hr_holidays_leaves_report_views.xml',
+        'report/hr_holidays_views.xml',
+    ],
+    'demo': [
+        'data/hr_holidays_demo.xml',
+    ],
     'qweb': [
         'static/src/xml/*.xml',
     ],
-    'test': ['test/test_hr_holiday.yml',
-             'test/hr_holidays_report.yml',
+    'test': [
+        'test/test_hr_holiday.yml',
+        'test/hr_holidays_report.yml',
     ],
     'installable': True,
     'application': True,
