+++ addons/hr_holidays/data/hr_holidays_demo.xml
