--- addons/hr_holidays/hr_bel_holidays_2008.xml
@@ -1,76 +1,0 @@
-<?xml version="1.0"?>
-<openerp>
-    <data>
-
-        <record model="hr.holidays">
-            <field name="name">New Year</field>
-            <field name="date_from">2008-01-01 08:00</field>
-            <field name="date_to">2008-01-01 18:00</field>
-            <field name="holiday_status" ref="hr.jf" />
-        </record>
-
-        <record model="hr.holidays">
-            <field name="name">Easter Monday</field>
-            <field name="date_from">2008-03-24 08:00</field>
-            <field name="date_to">2008-03-24 18:00</field>
-            <field name="holiday_status" ref="hr.jf" />
-        </record>
-
-        <record model="hr.holidays">
-            <field name="name">Labour Day</field>
-            <field name="date_from">2008-05-01 08:00</field>
-            <field name="date_to">2008-05-01 18:00</field>
-            <field name="holiday_status" ref="hr.jf" />
-        </record>
-
-        <record model="hr.holidays">
-            <field name="name">Ascension</field>
-            <field name="date_from">2008-05-02 08:00</field>
-            <field name="date_to">2008-05-02 18:00</field>
-            <field name="holiday_status" ref="hr.jf" />
-        </record>
-
-        <record model="hr.holidays">
-            <field name="name">Pentecost Monday</field>
-            <field name="date_from">2008-05-12 08:00</field>
-            <field name="date_to">2008-05-12 18:00</field>
-            <field name="holiday_status" ref="hr.jf" />
-        </record>
-
-        <record model="hr.holidays">
-            <field name="name">National Day</field>
-            <field name="date_from">2008-07-21 08:00</field>
-            <field name="date_to">2008-07-21 18:00</field>
-            <field name="holiday_status" ref="hr.jf" />
-        </record>
-
-        <record model="hr.holidays">
-            <field name="name">Assumption Day</field>
-            <field name="date_from">2008-08-15 08:00</field>
-            <field name="date_to">2008-08-15 18:00</field>
-            <field name="holiday_status" ref="hr.jf" />
-        </record>
-
-        <record model="hr.holidays">
-            <field name="name">All Saints</field>
-            <field name="date_from">2008-11-01 08:00</field>
-            <field name="date_to">2008-11-01 18:00</field>
-            <field name="holiday_status" ref="hr.jf" />
-        </record>
-
-        <record model="hr.holidays">
-            <field name="name">Armistice</field>
-            <field name="date_from">2008-11-11 08:00</field>
-            <field name="date_to">2008-11-11 18:00</field>
-            <field name="holiday_status" ref="hr.jf" />
-        </record>
-
-        <record model="hr.holidays">
-            <field name="name">Christmas</field>
-            <field name="date_from">2008-12-25 08:00</field>
-            <field name="date_to">2008-12-25 18:00</field>
-            <field name="holiday_status" ref="hr.jf" />
-        </record>
-
-    </data>
-</openerp>
