+++ addons/hr_holidays/report/hr_holidays_leaves_report_views.xml
