+++ addons/hr_holidays/report/hr_holidays_reports.xml
