+++ addons/hr_holidays/report/hr_holidays_templates.xml
