+++ addons/hr_holidays/security/hr_holidays_security.xml
