+++ addons/hr_holidays/views/hr_holidays_views.xml
