+++ addons/hr_holidays/views/hr_views.xml
