+++ addons/hr_holidays/wizard/hr_holidays_summary_department_views.xml
