+++ addons/hr_holidays/wizard/hr_holidays_summary_employees_views.xml
